src/requirements.txt
Authlib==0.15.3
bcrypt==3.2.0
cryptography==3.3.2
Flask==1.1.2
flask-talisman==0.7.0
flask-login==0.5.0
Flask-WTF==0.14.3
google-api-python-client==1.12.8
google-cloud-datastore==1.15.3
google-cloud-logging==2.2.0
google-cloud-ndb==1.7.2
google-cloud-storage==1.35.0
grpcio==1.34.0
gunicorn==20.0.4
passlib==1.7.4
reportlab==3.5.58
sendgrid==6.5.0
xhtml2pdf==0.2.5
wtforms[email]==2.3.3
